requirements.txt
autopep8==1.4
boto3==1.9.0
botocore==1.12.1
dj-database-url==0.4.2
Django==2.1.1
django-storages==1.7.1
django-tagging==0.4.6
docutils==0.14
gunicorn==19.7.1
jmespath==0.9.3
Pillow==5.2.0
psycopg2==2.7.1
pycodestyle==2.4.0
python-dateutil==2.7.3
pytz==2018.5
s3transfer==0.1.13
six==1.11.0
urllib3==1.23
whitenoise==3.3.0
